consolidate hooks
